feat: ERM Form
Split out ERM Form into a proper component, and leave the hook implementation as is. Cleans up implementation in our apps
